feat(uki): expose build-uki as pkg/uki.Build
Extracts the UKI build pipeline from internal/cmd/build-uki.go into a new
pkg/uki package so it can be consumed as a library (e.g. by daedalus)
instead of shelling out to the auroraboot binary. The CLI command is now
a thin wrapper that parses flags into uki.Options and calls uki.Build.
