+++ pkg/uki/suite_test.go
@@ -1,0 +1,13 @@
+package uki
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestUki(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "pkg/uki suite")
+}
