+++ pkg/uki/uki_test.go
@@ -1,4 +1,4 @@
-package cmd
+package uki
 
 import (
 	"os"
